Squirrel turret and enemy spawners

+++ Assets/Scripts/Managers/HealthManager.cs
@@ -19,7 +19,7 @@ public class HealthManager : MonoBehaviour
         return health;
     }
 
-    public void TakeDamage(int damage) {
+    public void TakeDamage(float damage) {
         health -= damage;
         if (health <= 0 && NoHealth != null) {
             enemyDead.Raise();
